+++ NurseManager/src/app/Sugestao/ranker.ts
@@ -1,14 +1,18 @@
-import { Enfermeiro } from './enfermeiro';
-import { Setor } from './setor';
+import { Enfermeiro } from '../model/enfermeiro';
+import { Setor } from '../model/setor';
+import { Turno } from '../model/turno';
 
 export class Ranker {
     
     getRanking(enfermeiro: Enfermeiro, setor: Setor, turno: string): number {
         var ranking: number = 0;
-        var titulacao: string[] = enfermeiro.especialidadeList;
-        var especializacao: string[] = enfermeiro.titulacaoList;
+        var titulacao: string[] = enfermeiro.titulacaoList;
+        var especializacao: string[];
+        for(let e of enfermeiro.especialidadeList){
+            especializacao.push(e.nome);
+        }
         //se o enfermeiro está disponível no turno, calcula seu ranking para o turno
-        if(enfermeiro.turno == turno){ 
+        if(enfermeiro.turno.nome == turno){ 
             for(let e of especializacao){
                 if(e == setor.especialidadeSetor){
                     ranking = ranking + 2;
@@ -16,9 +20,10 @@ export class Ranker {
             }
             for(let t of titulacao){
                 if(t == setor.especialidadeSetor){
-                    ranking = ranking + 1,5;
+                    ranking = ranking + 2;
                 }
             }
+            ranking = ranking + enfermeiro.experiencia;
         }
         return ranking;
       }
